backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-socketio==5.12.1
python-multipart==0.0.20
pydantic==2.11.7
httpx==0.28.1
